refactor(cli): move byte-output helper into cli and restore comment anchor
Project-standards review findings.

output_description_bytes lived in main.rs, which had grown back to 1309
lines -- larger than before the U6 extraction, because the review fixes
added to it. It is CLI output support rather than binary orchestration, so
it belongs with synthetic_result in src/cli/mod.rs. main.rs is now 1276.

cargo fmt had sorted the new `use cli::...` lines between the
"used only by the unix-gated path" comment and the #[cfg(unix)] import it
describes, leaving it attached to unconditionally-used imports. Restored.

Not applied: splitting tests/cli_integration.rs (1567 lines, over the
500-600 guidance). Integration tests are separate binaries and this repo
has no tests/common/ module, so a thematic split needs one introduced and
the pre-existing 645 lines refactored onto it -- a wider change than this
PR should carry. Recorded as a residual.

Also stale rather than actioned: the review flagged GOTCHAS S17.3 as
describing the old escape set, but it was read from a snapshot predating
the previous commit, which already rewrote that section.

Verified: 24 suites pass, clippy clean, and both differentials still match
file-5.41 byte-for-byte.
